requirements.txt
# Base requirements for running the OSF.
# NOTE: This does not include addon, development or release requirements.
# NOTE: When updating pinned version, you may also need to update constraints.txt
# To install addon requirements: inv requirements --addons
# To install dev requirements: inv requirements --dev
# To install release requirements: inv requirements --release

invoke==0.15.0
Werkzeug==0.10.4  # pyup: ignore # newer versions break local development
Flask==0.10.1
gevent==1.2.2
Mako==1.0.7
Markdown==2.6.9
WTForms==1.0.4
beautifulsoup4==4.6.0
# Fork of celery 4.2.1 with https://github.com/celery/celery/pull/4839 merged
git+https://github.com/providedh/celery@v4.2.1+cos0
kombu==4.2.0
amqp==2.2.2
vine==1.1.4
httplib2==0.10.3
itsdangerous==0.24
lxml==3.4.1
mailchimp==2.0.9
nameparser==0.5.3
bcrypt==3.1.4
python-dateutil==2.6.1
python-gnupg==0.4.1
pytz==2017.3
bleach==2.1.1
html5lib==0.999999999
blinker==1.4
furl==0.4.92
elasticsearch==2.4.0  # pyup: >=2.4,<3.0 # Major version must be same as ES version
google-api-python-client==1.6.4
Babel==2.5.1
citeproc-py==0.4.0
boto3==1.4.7
django-waffle==0.13.0
pymongo==3.6.0
progressbar==2.3
# Support SSL SNI on Python < 2.7.9
# (http://docs.python-requests.org/en/latest/community/faq/#what-are-hostname-doesn-t-match-errors)
ndg-httpsclient==0.3.0
# Python markdown extensions for comment emails
git+git://github.com/providedh/mdx_del_ins.git

# Issue: certifi-2015.9.6.1 and 2015.9.6.2 fail verification (https://github.com/certifi/python-certifi/issues/26)
# MailChimp Ticket: LTK1218902287135X, Domain: https://us9.api.mailchimp.com
certifi==2015.4.28
sendgrid==1.5.13

# avoid dbus-python requirement from secretstorage dependency
# https://github.com/jaraco/keyring/blob/9.1/setup.py
keyring==9.1

requests==2.18.4
urllib3==1.22
oauthlib==2.0.6
requests-oauthlib==0.8.0
raven==5.32.0

# API requirements
Django==1.11.13  # pyup: <2.0 # Remove this when we're on Py3
djangorestframework==3.6.3
django-cors-headers==1.3.1
djangorestframework-bulk==0.2.1
pyjwt==1.5.3
# django-celery-beat==1.0.1  # BSD 3 Clause
# Contains a fix for handling disabled tasks that still has not been release
git+git://github.com/celery/django-celery-beat@f014edcb954c707cb7628f4416257b6a58689523# BSD 3 Clause
# Issue: sorry, but this version only supports 100 named groups (https://github.com/eliben/pycparser/issues/147)
pycparser==2.18
pyjwe==1.0.0
jsonschema==2.6.0
django-guardian==1.4.9

# Admin requirements
django-webpack-loader==0.5.0
django-nose==1.4.5
django-password-reset==1.0
sendgrid-django==2.0.0

# Analytics requirements
keen==0.5.1
maxminddb-geolite2==2018.308

# OSF models
django-typed-models==0.7.0
django-storages==1.6.6
google-cloud-storage==0.22.0  # dependency of django-storages, hard-pin to version
git+https://github.com/providedh/django-dirtyfields@develop
git+https://github.com/providedh/django-extensions@master
django-include==0.2.4
psycopg2==2.7.3
ujson==1.35
sqlparse==0.2.4
psycogreen==1.0
django-bulk-update==2.2.0
ciso8601==1.0.5

# Reviews requirements
transitions==0.6.1
enum34==1.1.6

# identifiers
datacite==1.0.1

# test parameterizing requirement
parameterized==0.7.0

# websockets in tei close reading annotation requirements
channels==1.1.8
asgi-rabbitmq==0.5.5
django-channels-presence==0.0.9
